Fixing timing issues for tests
